+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## v4.1.1 - ?
+- Ported ip plugin to std
 - Improve error logging in std::Package
 
 ## v4.1.0 - 2023-01-13